+++ sources/Google.Solutions.IapDesktop.Application/Theme/Themes.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright 2024 Google LLC
+// Copyright 2020 Google LLC
 //
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
@@ -19,7 +19,11 @@
 // under the License.
 //
 
+using Google.Solutions.Common.Util;
+using Google.Solutions.IapDesktop.Application.Profile.Settings;
 using Google.Solutions.Mvvm.Theme;
+using Google.Solutions.Settings.Collection;
+using System.Windows.Forms;
 using WeifenLuo.WinFormsUI.Docking;
 
 namespace Google.Solutions.IapDesktop.Application.Theme
@@ -48,5 +52,152 @@ namespace Google.Solutions.IapDesktop.Application.Theme
         /// Theme for the docking suite.
         /// </summary>
         ThemeBase DockPanelTheme { get; }
+    }
+
+    public class Themes
+    {
+        /// <summary>
+        /// Theme for system dialogs.
+        /// </summary>
+        public ISystemDialogTheme SystemDialog{ get; }
+
+        /// <summary>
+        /// Theme for dialogs.
+        /// </summary>
+        public IDialogTheme Dialog { get; }
+
+        /// <summary>
+        /// Theme for tool windows.
+        /// </summary>
+        public IToolWindowTheme ToolWindow { get; }
+
+        /// <summary>
+        /// Theme for the main window.
+        /// </summary>
+        public IMainWindowTheme MainWindow { get; }
+
+        private Themes(
+            ISystemDialogTheme systemDialogTheme,
+            IDialogTheme dialogTheme,
+            IToolWindowTheme toolWindowTheme,
+            IMainWindowTheme mainWindowTheme)
+        {
+            this.SystemDialog = systemDialogTheme;
+            this.Dialog = dialogTheme;
+            this.ToolWindow = toolWindowTheme;
+            this.MainWindow = mainWindowTheme;
+        }
+
+        /// <summary>
+        /// Load themes from the respository.
+        /// </summary>
+        public static Themes Load(IRepository<IThemeSettings> themeSettingsRepository)
+        {
+            var settings = themeSettingsRepository.GetSettings();
+            var windowsTheme = settings.Theme.Value switch
+            {
+                //
+                // Use same mode as Windows.
+                //
+                ApplicationTheme.System => new WindowsRuleSet(SystemTheme.ShouldAppsUseDarkMode),
+
+                //
+                // Use dark mode if possible.
+                //
+                ApplicationTheme.Dark => new WindowsRuleSet(SystemTheme.IsDarkModeSupported),
+
+                //
+                // Use safe defaults that also work on downlevel
+                // versions of Windows.
+                //
+                _ => new WindowsRuleSet(false),
+            };
+
+            var vsTheme = windowsTheme.IsDarkModeEnabled
+                ? VSTheme.GetDarkTheme()
+                : VSTheme.GetLightTheme();
+
+            var systemDialogTheme = new ControlTheme()
+                .AddRuleSet(windowsTheme)
+                .AddRuleSet(new WindowsSystemDialogRuleset())
+                .AddRuleSet(new VSThemeDialogRuleSet(vsTheme))
+                .AddRuleSet(new GdiScalingRuleset())
+                .AddRuleSet(new DpiAwarenessRuleset());
+
+            var dialogTheme = new ControlTheme()
+                .AddRuleSet(windowsTheme)
+                .AddRuleSet(new CommonControlRuleSet())
+                .AddRuleSet(new VSThemeDialogRuleSet(vsTheme))
+                .AddRuleSet(new GdiScalingRuleset())
+                .AddRuleSet(new DpiAwarenessRuleset());
+
+            var dockWindowTheme = new ControlTheme()
+                .AddRuleSet(windowsTheme)
+                .AddRuleSet(new CommonControlRuleSet())
+                .AddRuleSet(new VSThemeDockWindowRuleSet(vsTheme))
+                .AddRuleSet(new GdiScalingRuleset())
+                .AddRuleSet(new DpiAwarenessRuleset());
+
+
+            if (vsTheme.Extender.FloatWindowFactory is VSThemeExtensions.FloatWindowFactory factory)
+            {
+                factory.Theme = dialogTheme;
+            }
+
+            return new Themes(
+                new SystemDialogTheme(systemDialogTheme),
+                new DialogWindowTheme(dialogTheme),
+                new ToolWindowTheme(dockWindowTheme),
+                new MainWindowTheme(dockWindowTheme, vsTheme));
+        }
+
+        //---------------------------------------------------------------------
+        // Typed theme wrappers.
+        //---------------------------------------------------------------------
+
+        private abstract class WindowThemeBase : IControlTheme
+        {
+            private readonly IControlTheme theme;
+
+            protected WindowThemeBase(IControlTheme theme)
+            {
+                this.theme = theme.ExpectNotNull(nameof(theme));
+            }
+
+            public void ApplyTo(Control control)
+            {
+                this.theme.ApplyTo(control);
+            }
+        }
+
+        private class SystemDialogTheme : WindowThemeBase, ISystemDialogTheme
+        {
+            internal SystemDialogTheme(IControlTheme theme) : base(theme)
+            { }
+        }
+
+        private class DialogWindowTheme : WindowThemeBase, IDialogTheme
+        {
+            internal DialogWindowTheme(IControlTheme theme) : base(theme)
+            { }
+        }
+
+        private class MainWindowTheme : WindowThemeBase, IMainWindowTheme
+        {
+            internal MainWindowTheme(
+                IControlTheme theme,
+                ThemeBase dockPanelTheme) : base(theme)
+            {
+                this.DockPanelTheme = dockPanelTheme;
+            }
+
+            public ThemeBase DockPanelTheme { get;}
+        }
+
+        private class ToolWindowTheme : WindowThemeBase, IToolWindowTheme
+        {
+            internal ToolWindowTheme(IControlTheme theme) : base(theme)
+            { }
+        }
     }
 }
